requirements.txt
deepdiff
PyYAML
selenium
Pygments